API refactor.
