+++ lib/iris/fileformats/netcdf.py
@@ -843,9 +843,9 @@ def load_cubes(filenames, callback=None, constraints=None):
     # TODO: rationalise UGRID/mesh handling once experimental.ugrid is folded
     #  into standard behaviour.
     # Deferred import to avoid circular imports.
-    from iris.experimental.ugrid import (
+    from iris.experimental.ugrid.cf import CFUGridReader
+    from iris.experimental.ugrid.load import (
         PARSE_UGRID_ON_LOAD,
-        CFUGridReader,
         _build_mesh_coords,
         _meshes_from_cf,
     )
@@ -2077,7 +2077,8 @@ class Saver:
                     # Not a cube coord, so must be a connectivity or
                     # element-coordinate of the mesh.
                     # Name it for it's first dim, i.e. mesh-dim of its location.
-                    from iris.experimental.ugrid import Connectivity
+
+                    from iris.experimental.ugrid.mesh import Connectivity
 
                     # At present, a location-coord cannot be nameless, as the
                     # Mesh code relies on guess_coord_axis.
