+++ lib/iris/tests/integration/experimental/test_ugrid_load.py
@@ -17,13 +17,13 @@ import iris.tests as tests  # isort:skip
 from collections.abc import Iterable
 
 from iris import Constraint, load
-from iris.experimental.ugrid import (
+from iris.experimental.ugrid import logger
+from iris.experimental.ugrid.load import (
     PARSE_UGRID_ON_LOAD,
-    Mesh,
     load_mesh,
     load_meshes,
-    logger,
 )
+from iris.experimental.ugrid.mesh import Mesh
 from iris.tests.stock.netcdf import (
     _file_from_cdl_template as create_file_from_cdl_template,
 )
