+++ lib/iris/tests/unit/common/metadata/test_metadata_manager_factory.py
@@ -25,7 +25,7 @@ from iris.common.metadata import (
     CubeMetadata,
     metadata_manager_factory,
 )
-from iris.experimental.ugrid import ConnectivityMetadata
+from iris.experimental.ugrid.metadata import ConnectivityMetadata
 
 BASES = [
     AncillaryVariableMetadata,
