+++ lib/iris/tests/unit/common/mixin/test_CFVariableMixin.py
@@ -25,7 +25,7 @@ from iris.common.metadata import (
     CubeMetadata,
 )
 from iris.common.mixin import CFVariableMixin, LimitedAttributeDict
-from iris.experimental.ugrid import ConnectivityMetadata
+from iris.experimental.ugrid.metadata import ConnectivityMetadata
 
 
 class Test__getter(tests.IrisTest):
