+++ lib/iris/tests/unit/experimental/ugrid/cf/test_CFUGridAuxiliaryCoordinateVariable.py
@@ -10,14 +10,15 @@ todo: fold these tests into cf tests when experimental.ugrid is folded into
  standard behaviour.
 
 """
-import numpy as np
-
-from iris.experimental.ugrid import CFUGridAuxiliaryCoordinateVariable, logger
-
 # Import iris.tests first so that some things can be initialised before
 # importing anything else.
-import iris.tests as tests
-from iris.tests.unit.experimental.ugrid.test_CFUGridReader import (
+import iris.tests as tests  # isort:skip
+
+import numpy as np
+
+from iris.experimental.ugrid import logger
+from iris.experimental.ugrid.cf import CFUGridAuxiliaryCoordinateVariable
+from iris.tests.unit.experimental.ugrid.cf.test_CFUGridReader import (
     netcdf_ugrid_variable,
 )
