+++ lib/iris/tests/unit/experimental/ugrid/cf/test_CFUGridConnectivityVariable.py
@@ -10,18 +10,16 @@ todo: fold these tests into cf tests when experimental.ugrid is folded into
  standard behaviour.
 
 """
-import numpy as np
-
-from iris.experimental.ugrid import (
-    CFUGridConnectivityVariable,
-    Connectivity,
-    logger,
-)
-
 # Import iris.tests first so that some things can be initialised before
 # importing anything else.
-import iris.tests as tests
-from iris.tests.unit.experimental.ugrid.test_CFUGridReader import (
+import iris.tests as tests  # isort:skip
+
+import numpy as np
+
+from iris.experimental.ugrid import logger
+from iris.experimental.ugrid.cf import CFUGridConnectivityVariable
+from iris.experimental.ugrid.mesh import Connectivity
+from iris.tests.unit.experimental.ugrid.cf.test_CFUGridReader import (
     netcdf_ugrid_variable,
 )
