+++ lib/iris/tests/unit/experimental/ugrid/cf/test_CFUGridGroup.py
@@ -10,19 +10,19 @@ todo: fold these tests into cf tests when experimental.ugrid is folded into
  standard behaviour.
 
 """
+# Import iris.tests first so that some things can be initialised before
+# importing anything else.
+import iris.tests as tests  # isort:skip
+
 from unittest.mock import MagicMock
 
-from iris.experimental.ugrid import (
+from iris.experimental.ugrid.cf import (
     CFUGridAuxiliaryCoordinateVariable,
     CFUGridConnectivityVariable,
     CFUGridGroup,
     CFUGridMeshVariable,
 )
 from iris.fileformats.cf import CFCoordinateVariable, CFDataVariable
-
-# Import iris.tests first so that some things can be initialised before
-# importing anything else.
-import iris.tests as tests
 
 
 class Tests(tests.IrisTest):
