+++ lib/iris/tests/unit/experimental/ugrid/load/test_ParseUgridOnLoad.py
@@ -9,11 +9,11 @@ Unit tests for the :class:`iris.experimental.ugrid.ParseUgridOnLoad` class.
 todo: remove this module when experimental.ugrid is folded into standard behaviour.
 
 """
-from iris.experimental.ugrid import PARSE_UGRID_ON_LOAD, ParseUGridOnLoad
-
 # Import iris.tests first so that some things can be initialised before
 # importing anything else.
-import iris.tests as tests
+import iris.tests as tests  # isort:skip
+
+from iris.experimental.ugrid.load import PARSE_UGRID_ON_LOAD, ParseUGridOnLoad
 
 
 class TestClass(tests.IrisTest):
