+++ lib/iris/tests/unit/experimental/ugrid/mesh/test_Connectivity.py
@@ -5,17 +5,17 @@
 # licensing details.
 """Unit tests for the :class:`iris.experimental.ugrid.Connectivity` class."""
 
+# Import iris.tests first so that some things can be initialised before
+# importing anything else.
+import iris.tests as tests  # isort:skip
+
 from xml.dom import minidom
 
 import numpy as np
 from numpy import ma
 
 from iris._lazy_data import as_lazy_data, is_lazy_data
-from iris.experimental.ugrid import Connectivity
-
-# Import iris.tests first so that some things can be initialised before
-# importing anything else.
-import iris.tests as tests
+from iris.experimental.ugrid.mesh import Connectivity
 
 
 class TestStandard(tests.IrisTest):
