+++ lib/iris/tests/unit/experimental/ugrid/metadata/test_ConnectivityMetadata.py
@@ -4,9 +4,12 @@
 # See COPYING and COPYING.LESSER in the root of the repository for full
 # licensing details.
 """
-Unit tests for the :class:`iris.experimental.ugrid.MeshCoordMetadata`.
+Unit tests for the :class:`iris.experimental.ugrid.ConnectivityMetadata`.
 
 """
+# Import iris.tests first so that some things can be initialised before
+# importing anything else.
+import iris.tests as tests  # isort:skip
 
 from copy import deepcopy
 import unittest.mock as mock
@@ -14,23 +17,20 @@ from unittest.mock import sentinel
 
 from iris.common.lenient import _LENIENT, _qualname
 from iris.common.metadata import BaseMetadata
-from iris.experimental.ugrid import MeshCoordMetadata
-
-# Import iris.tests first so that some things can be initialised before
-# importing anything else.
-import iris.tests as tests
+from iris.experimental.ugrid.metadata import ConnectivityMetadata
 
 
-class Test__identity(tests.IrisTest):
+class Test(tests.IrisTest):
     def setUp(self):
         self.standard_name = mock.sentinel.standard_name
         self.long_name = mock.sentinel.long_name
         self.var_name = mock.sentinel.var_name
         self.units = mock.sentinel.units
         self.attributes = mock.sentinel.attributes
-        self.location = mock.sentinel.location
-        self.axis = mock.sentinel.axis
-        self.cls = MeshCoordMetadata
+        self.cf_role = mock.sentinel.cf_role
+        self.start_index = mock.sentinel.start_index
+        self.src_dim = mock.sentinel.src_dim
+        self.cls = ConnectivityMetadata
 
     def test_repr(self):
         metadata = self.cls(
@@ -39,13 +39,14 @@ class Test__identity(tests.IrisTest):
             var_name=self.var_name,
             units=self.units,
             attributes=self.attributes,
-            location=self.location,
-            axis=self.axis,
+            cf_role=self.cf_role,
+            start_index=self.start_index,
+            src_dim=self.src_dim,
         )
         fmt = (
-            "MeshCoordMetadata(standard_name={!r}, long_name={!r}, "
-            "var_name={!r}, units={!r}, attributes={!r}, "
-            "location={!r}, axis={!r})"
+            "ConnectivityMetadata(standard_name={!r}, long_name={!r}, "
+            "var_name={!r}, units={!r}, attributes={!r}, cf_role={!r}, "
+            "start_index={!r}, src_dim={!r})"
         )
         expected = fmt.format(
             self.standard_name,
@@ -53,8 +54,9 @@ class Test__identity(tests.IrisTest):
             self.var_name,
             self.units,
             self.attributes,
-            self.location,
-            self.axis,
+            self.cf_role,
+            self.start_index,
+            self.src_dim,
         )
         self.assertEqual(expected, repr(metadata))
 
@@ -65,8 +67,9 @@ class Test__identity(tests.IrisTest):
             "var_name",
             "units",
             "attributes",
-            "location",
-            "axis",
+            "cf_role",
+            "start_index",
+            "src_dim",
         )
         self.assertEqual(self.cls._fields, expected)
 
@@ -82,11 +85,17 @@ class Test__eq__(tests.IrisTest):
             var_name=sentinel.var_name,
             units=sentinel.units,
             attributes=sentinel.attributes,
-            location=sentinel.location,
-            axis=sentinel.axis,
+            cf_role=sentinel.cf_role,
+            start_index=sentinel.start_index,
+            src_dim=sentinel.src_dim,
         )
         self.dummy = sentinel.dummy
-        self.cls = MeshCoordMetadata
+        self.cls = ConnectivityMetadata
+        # The "src_dim" member is stateful only, and does not participate in
+        # lenient/strict equivalence.
+        self.members_no_src_dim = filter(
+            lambda member: member != "src_dim", self.cls._members
+        )
 
     def test_wraps_docstring(self):
         self.assertEqual(BaseMetadata.__eq__.__doc__, self.cls.__eq__.__doc__)
@@ -120,10 +129,10 @@ class Test__eq__(tests.IrisTest):
             self.assertTrue(lmetadata.__eq__(rmetadata))
             self.assertTrue(rmetadata.__eq__(lmetadata))
 
-    def test_op_lenient_same_none_nonmember(self):
+    def test_op_lenient_same_none(self):
         lmetadata = self.cls(**self.values)
         right = self.values.copy()
-        right["long_name"] = None
+        right["var_name"] = None
         rmetadata = self.cls(**right)
 
         with mock.patch("iris.common.metadata._LENIENT", return_value=True):
@@ -131,7 +140,7 @@ class Test__eq__(tests.IrisTest):
             self.assertTrue(rmetadata.__eq__(lmetadata))
 
     def test_op_lenient_same_members_none(self):
-        for member in self.cls._members:
+        for member in self.members_no_src_dim:
             lmetadata = self.cls(**self.values)
             right = self.values.copy()
             right[member] = None
@@ -143,10 +152,20 @@ class Test__eq__(tests.IrisTest):
                 self.assertFalse(lmetadata.__eq__(rmetadata))
                 self.assertFalse(rmetadata.__eq__(lmetadata))
 
+    def test_op_lenient_same_src_dim_none(self):
+        lmetadata = self.cls(**self.values)
+        right = self.values.copy()
+        right["src_dim"] = None
+        rmetadata = self.cls(**right)
+
+        with mock.patch("iris.common.metadata._LENIENT", return_value=True):
+            self.assertTrue(lmetadata.__eq__(rmetadata))
+            self.assertTrue(rmetadata.__eq__(lmetadata))
+
     def test_op_lenient_different(self):
         lmetadata = self.cls(**self.values)
         right = self.values.copy()
-        right["long_name"] = self.dummy
+        right["units"] = self.dummy
         rmetadata = self.cls(**right)
 
         with mock.patch("iris.common.metadata._LENIENT", return_value=True):
@@ -154,7 +173,7 @@ class Test__eq__(tests.IrisTest):
             self.assertFalse(rmetadata.__eq__(lmetadata))
 
     def test_op_lenient_different_members(self):
-        for member in self.cls._members:
+        for member in self.members_no_src_dim:
             lmetadata = self.cls(**self.values)
             right = self.values.copy()
             right[member] = self.dummy
@@ -165,6 +184,16 @@ class Test__eq__(tests.IrisTest):
             ):
                 self.assertFalse(lmetadata.__eq__(rmetadata))
                 self.assertFalse(rmetadata.__eq__(lmetadata))
+
+    def test_op_lenient_different_src_dim(self):
+        lmetadata = self.cls(**self.values)
+        right = self.values.copy()
+        right["src_dim"] = self.dummy
+        rmetadata = self.cls(**right)
+
+        with mock.patch("iris.common.metadata._LENIENT", return_value=True):
+            self.assertTrue(lmetadata.__eq__(rmetadata))
+            self.assertTrue(rmetadata.__eq__(lmetadata))
 
     def test_op_strict_same(self):
         lmetadata = self.cls(**self.values)
@@ -185,7 +214,7 @@ class Test__eq__(tests.IrisTest):
             self.assertFalse(rmetadata.__eq__(lmetadata))
 
     def test_op_strict_different_members(self):
-        for member in self.cls._members:
+        for member in self.members_no_src_dim:
             lmetadata = self.cls(**self.values)
             right = self.values.copy()
             right[member] = self.dummy
@@ -196,6 +225,16 @@ class Test__eq__(tests.IrisTest):
             ):
                 self.assertFalse(lmetadata.__eq__(rmetadata))
                 self.assertFalse(rmetadata.__eq__(lmetadata))
+
+    def test_op_strict_different_src_dim(self):
+        lmetadata = self.cls(**self.values)
+        right = self.values.copy()
+        right["src_dim"] = self.dummy
+        rmetadata = self.cls(**right)
+
+        with mock.patch("iris.common.metadata._LENIENT", return_value=False):
+            self.assertTrue(lmetadata.__eq__(rmetadata))
+            self.assertTrue(rmetadata.__eq__(lmetadata))
 
     def test_op_strict_different_none(self):
         lmetadata = self.cls(**self.values)
@@ -208,7 +247,7 @@ class Test__eq__(tests.IrisTest):
             self.assertFalse(rmetadata.__eq__(lmetadata))
 
     def test_op_strict_different_members_none(self):
-        for member in self.cls._members:
+        for member in self.members_no_src_dim:
             lmetadata = self.cls(**self.values)
             right = self.values.copy()
             right[member] = None
@@ -220,24 +259,24 @@ class Test__eq__(tests.IrisTest):
                 self.assertFalse(lmetadata.__eq__(rmetadata))
                 self.assertFalse(rmetadata.__eq__(lmetadata))
 
+    def test_op_strict_different_src_dim_none(self):
+        lmetadata = self.cls(**self.values)
+        right = self.values.copy()
+        right["src_dim"] = None
+        rmetadata = self.cls(**right)
+
+        with mock.patch("iris.common.metadata._LENIENT", return_value=False):
+            self.assertTrue(lmetadata.__eq__(rmetadata))
+            self.assertTrue(rmetadata.__eq__(lmetadata))
+
 
 class Test___lt__(tests.IrisTest):
     def setUp(self):
-        self.cls = MeshCoordMetadata
-        values = [1] * len(self.cls._fields)
-        self.one = self.cls(*values)
-
-        values_two = values[:]
-        values_two[2] = 2
-        self.two = self.cls(*values_two)
-
-        values_none = values[:]
-        values_none[2] = None
-        self.none = self.cls(*values_none)
-
-        values_attrs = values[:]
-        values_attrs[4] = 10
-        self.attributes = self.cls(*values_attrs)
+        self.cls = ConnectivityMetadata
+        self.one = self.cls(1, 1, 1, 1, 1, 1, 1, 1)
+        self.two = self.cls(1, 1, 1, 2, 1, 1, 1, 1)
+        self.none = self.cls(1, 1, 1, None, 1, 1, 1, 1)
+        self.attributes = self.cls(1, 1, 1, 1, 10, 1, 1, 1)
 
     def test__ascending_lt(self):
         result = self.one < self.two
@@ -264,17 +303,18 @@ class Test___lt__(tests.IrisTest):
 
 class Test_combine(tests.IrisTest):
     def setUp(self):
-        self.cls = MeshCoordMetadata
         self.values = dict(
             standard_name=sentinel.standard_name,
             long_name=sentinel.long_name,
             var_name=sentinel.var_name,
             units=sentinel.units,
             attributes=sentinel.attributes,
-            location=sentinel.location,
-            axis=sentinel.axis,
+            cf_role=sentinel.cf_role,
+            start_index=sentinel.start_index,
+            src_dim=sentinel.src_dim,
         )
         self.dummy = sentinel.dummy
+        self.cls = ConnectivityMetadata
         self.none = self.cls(*(None,) * len(self.cls._fields))
 
     def test_wraps_docstring(self):
@@ -329,7 +369,7 @@ class Test_combine(tests.IrisTest):
     def test_op_lenient_same_none(self):
         lmetadata = self.cls(**self.values)
         right = self.values.copy()
-        right["long_name"] = None
+        right["var_name"] = None
         rmetadata = self.cls(**right)
         expected = self.values
 
@@ -358,10 +398,10 @@ class Test_combine(tests.IrisTest):
     def test_op_lenient_different(self):
         lmetadata = self.cls(**self.values)
         right = self.values.copy()
-        right["long_name"] = self.dummy
+        right["units"] = self.dummy
         rmetadata = self.cls(**right)
         expected = self.values.copy()
-        expected["long_name"] = None
+        expected["units"] = None
 
         with mock.patch("iris.common.metadata._LENIENT", return_value=True):
             self.assertEqual(expected, lmetadata.combine(rmetadata)._asdict())
@@ -460,17 +500,18 @@ class Test_combine(tests.IrisTest):
 
 class Test_difference(tests.IrisTest):
     def setUp(self):
-        self.cls = MeshCoordMetadata
         self.values = dict(
             standard_name=sentinel.standard_name,
             long_name=sentinel.long_name,
             var_name=sentinel.var_name,
             units=sentinel.units,
             attributes=sentinel.attributes,
-            location=sentinel.location,
-            axis=sentinel.axis,
+            cf_role=sentinel.cf_role,
+            start_index=sentinel.start_index,
+            src_dim=sentinel.src_dim,
         )
         self.dummy = sentinel.dummy
+        self.cls = ConnectivityMetadata
         self.none = self.cls(*(None,) * len(self.cls._fields))
 
     def test_wraps_docstring(self):
@@ -524,7 +565,7 @@ class Test_difference(tests.IrisTest):
     def test_op_lenient_same_none(self):
         lmetadata = self.cls(**self.values)
         right = self.values.copy()
-        right["long_name"] = None
+        right["var_name"] = None
         rmetadata = self.cls(**right)
 
         with mock.patch("iris.common.metadata._LENIENT", return_value=True):
@@ -557,12 +598,12 @@ class Test_difference(tests.IrisTest):
         left = self.values.copy()
         lmetadata = self.cls(**left)
         right = self.values.copy()
-        right["long_name"] = self.dummy
+        right["units"] = self.dummy
         rmetadata = self.cls(**right)
         lexpected = deepcopy(self.none)._asdict()
-        lexpected["long_name"] = (left["long_name"], right["long_name"])
+        lexpected["units"] = (left["units"], right["units"])
         rexpected = deepcopy(self.none)._asdict()
-        rexpected["long_name"] = lexpected["long_name"][::-1]
+        rexpected["units"] = lexpected["units"][::-1]
 
         with mock.patch("iris.common.metadata._LENIENT", return_value=True):
             self.assertEqual(
@@ -687,7 +728,7 @@ class Test_difference(tests.IrisTest):
 
 class Test_equal(tests.IrisTest):
     def setUp(self):
-        self.cls = MeshCoordMetadata
+        self.cls = ConnectivityMetadata
         self.none = self.cls(*(None,) * len(self.cls._fields))
 
     def test_wraps_docstring(self):
