+++ lib/iris/tests/unit/experimental/ugrid/metadata/test_MeshMetadata.py
@@ -7,6 +7,9 @@
 Unit tests for the :class:`iris.experimental.ugrid.MeshMetadata`.
 
 """
+# Import iris.tests first so that some things can be initialised before
+# importing anything else.
+import iris.tests as tests  # isort:skip
 
 from copy import deepcopy
 import unittest.mock as mock
@@ -14,11 +17,7 @@ from unittest.mock import sentinel
 
 from iris.common.lenient import _LENIENT, _qualname
 from iris.common.metadata import BaseMetadata
-from iris.experimental.ugrid import MeshMetadata
-
-# Import iris.tests first so that some things can be initialised before
-# importing anything else.
-import iris.tests as tests
+from iris.experimental.ugrid.metadata import MeshMetadata
 
 
 class Test(tests.IrisTest):
